backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
django-cors-headers
python-dotenv
djangorestframework-simplejwt

# AI/ML Dependencies
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.3.0
python-dateutil>=2.8.0

# Additional utilities
celery>=5.3.0
redis>=4.5.0
